iot-api-discovery/requirements.txt
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
requests>=2.32.3
sqlalchemy>=2.0.32
pydantic>=2.8.2
python-dotenv>=1.0.1
fastapi>=0.114.0
uvicorn[standard]>=0.30.6
python-nmap>=0.7.1
paho-mqtt>=2.1.0
aiocoap>=0.4.7
pytest>=8.3.2
pytest-asyncio>=0.23.8
PyYAML>=6.0.2
readability-lxml>=0.8.1
zeroconf>=0.132.2
bleak>=0.22.3
pyzbar>=0.1.9
pillow>=10.4.0
sounddevice>=0.4.7
boto3>=1.34.159
psycopg2-binary>=2.9.9
alembic>=1.13.2
websockets>=12.0
